requirements.txt
PyGithub
launchpadlib
